Move basic controls to ControlBase.py
